move common components
